new features

+++ src/main/java/com/vats/projects/employee/management/system/dto/RoleDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class RoleDTO {
-    private Long id;
-    private String name; // e.g., ROLE_ADMIN, ROLE_MANAGER, ROLE_EMPLOYEE
+    private Integer id;
+    private String name; // e.g., ADMIN, MANAGER, EMPLOYEE
 }
